# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KhelSetu API
  version: 1.0.0
  description: >
    # KhelSetu API Documentation


    ## Platform Overview


    KhelSetu is a multi-tenant sports tournament management platform with **two
    distinct portals**:


    ### 🌐 Public Portal (No Login Required)

    The public portal is for **viewers and fans** who want to watch live
    matches, browse tournaments, and view standings. These endpoints are
    **completely open** — no authentication needed.


    - **Live match scores** — Real-time scores for any match

    - **Tournament browsing** — View all public tournaments

    - **Standings & fixtures** — League tables and match schedules

    - **Team & player profiles** — Roster and player information

    - **Global search** — Search across tournaments, teams, and players


    ### 🔒 SaaS Dashboard (Login Required)

    The dashboard is for **organization members** (admins, scorers, coaches) who
    manage tournaments, teams, players, and live scoring. **All dashboard
    endpoints require JWT authentication** and **organization membership**.


    - **Organization management** — Create/manage organizations, invite members

    - **Tournament management** — Create tournaments, generate fixtures

    - **Team & player management** — Roster management, player profiles

    - **Live scoring** — Real-time event-driven scoring (Football, Cricket,
    Basketball)

    - **Standings** — Auto-calculated league tables

...

...

...


...


...

...
    refresh token (7d)
...
paths:
...
  /api/billing/subscriptions:
    get:
      summary: Get organization subscription
      tags:
        - Billing
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription details
    post:
      summary: Create subscription
      tags:
        - Billing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - organization_id
                - plan_id
              properties:
                organization_id:
                  type: string
                plan_id:
                  type: string
                payment_method_id:
                  type: string
      responses:
        "201":
          description: Subscription created
  /api/billing/subscriptions/{id}:
    put:
      summary: Update subscription
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription updated
    delete:
      summary: Cancel subscription
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription cancelled
  /api/billing/invoices:
    get:
      summary: Get organization invoices
      tags:
        - Billing
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - paid
              - failed
              - cancelled
      responses:
        "200":
          description: List of invoices
  /api/billing/invoices/{id}:
    get:
      summary: Get invoice details
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Invoice details
  /api/billing/usage:
    get:
      summary: Get usage statistics
      tags:
        - Billing
      security:
        - bearerAuth: []
      parameters:
        - name: organization_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Usage data
  /api/billing/payment-methods:
    get:
      summary: Get payment methods
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of payment methods
    post:
      summary: Add payment method
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Payment method added
  /api/billing/payment-methods/{id}:
    delete:
      summary: Remove payment method
      tags:
        - Billing
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Payment method removed
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Get it from POST /api/auth/login. Pass as:
        Authorization: Bearer <token>"
...
security:
  - bearerAuth: []
...